feat: support ELK layout and interactive controls
BREAKING CHANGE: requires javascript module import
