+++ exampleSite/config/_default/server.toml
@@ -1,0 +1,28 @@
+# Auto-generated file - do not modify
+
+[[headers]]
+  for = '/**'
+
+  [headers.values]
+    Access-Control-Allow-Origin = '*'
+    Content-Security-Policy = """
+        base-uri 'self'; \
+        connect-src 'self'; \
+        default-src 'none'; \
+        font-src 'self'; \
+        form-action 'self'; \
+        img-src 'self' data:; \
+        manifest-src 'self'; \
+        media-src 'self'; \
+        object-src 'none'; \
+        script-src 'self'; \
+        style-src 'self' 'unsafe-inline'; \
+        """
+    Permissions-Policy = 'geolocation=(), midi=(), sync-xhr=(), microphone=(), camera=(), magnetometer=(), gyroscope=(), fullscreen=(), payment=() '
+    Referrer-Policy = 'strict-origin'
+    Strict-Transport-Security = 'max-age=31536000; includeSubDomains; preload'
+    X-Content-Type-Options = 'nosniff'
+    X-Frame-Options = 'SAMEORIGIN'
+    X-XSS-Protection = '1; mode=block'
+    cache-control = 'max-age=0, no-cache, no-store, must-revalidate '
+
