+++ exampleSite/go.mod
@@ -3,20 +3,7 @@ module github.com/gethinode/mod-mermaid/exampleSite
 go 1.19
 
 require (
-	github.com/airbnb/lottie-web v5.12.2+incompatible // indirect
-	github.com/gethinode/hinode v0.29.4 // indirect
-	github.com/gethinode/mod-bootstrap v1.3.3 // indirect
-	github.com/gethinode/mod-csp v1.0.7 // indirect
-	github.com/gethinode/mod-flexsearch/v2 v2.1.1 // indirect
-	github.com/gethinode/mod-fontawesome v1.10.1 // indirect
-	github.com/gethinode/mod-google-analytics v1.3.1 // indirect
-	github.com/gethinode/mod-katex v1.1.3 // indirect
-	github.com/gethinode/mod-leaflet v1.2.1 // indirect
-	github.com/gethinode/mod-lottie v1.5.14 // indirect
-	github.com/gethinode/mod-mermaid v1.0.0 // indirect
+	github.com/gethinode/mod-fontawesome/v2 v2.1.2 // indirect
 	github.com/gethinode/mod-mermaid/v2 v2.0.3 // indirect
-	github.com/gethinode/mod-simple-datatables v1.1.7 // indirect
-	github.com/gethinode/mod-utils/v4 v4.11.1 // indirect
-	github.com/nextapps-de/flexsearch v0.0.0-20250521173815-f12b83ec5c58 // indirect
-	github.com/twbs/bootstrap v5.3.6+incompatible // indirect
+	github.com/gethinode/mod-utils/v4 v4.12.0 // indirect
 )
