+++ exampleSite/hugo.toml
@@ -5,25 +5,21 @@ title = 'Test site for mod-mermaid'
 [module]
   # Build and serve using local mod-mermaid clone declared in the named Hugo workspace:
   workspace = "mod-mermaid.work"
-  [[module.imports]]
-    path = "github.com/gethinode/mod-mermaid/v2"
   [[module.mounts]]
     source = 'assets'
     target = 'assets'
+  [[module.mounts]]
+    source = 'data'
+    target = 'data'
   [[module.mounts]]
     source = 'layouts'
     target = 'layouts'
   [[module.mounts]]
     source = 'static'
     target = 'static'
-  [[module.imports.mounts]]
-    source = "layouts"
-    target = "layouts"
-  [[module.imports.mounts]]
-    source = "assets/js/modules/mermaid/mermaid-init.js"
-    target = "static/js/mermaid-init.js"
-  [[module.imports.mounts]]
-    source = "dist/mermaid.js"
-    target = "static/js/mermaid.js"
+  [[module.imports]]
+    path = "github.com/gethinode/mod-fontawesome/v2"
+  [[module.imports]]
+    path = "github.com/gethinode/mod-mermaid/v3"
   [[module.imports]]
     path = "github.com/gethinode/mod-utils/v4"
