+++ exampleSite/layouts/baseof.html
@@ -4,12 +4,20 @@
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0">
   <title>{{ with .Title }}{{ printf "%s | " . }}{{ end }}{{ site.Title }}</title>
-  <link rel=stylesheet href="style.css">
+  {{ partial "head/stylesheet.html" -}}
 </head>
 <body>
-{{ block "main" . }}
-{{ end }}
+  Settings:
+  <ul>
+    <li>elk = {{ site.Params.modules.mermaid.elk | default false }}</li>
+    <li>look = {{ site.Params.modules.mermaid.look | default "classic" }}</li>
+    <li>layout = {{ site.Params.modules.mermaid.layout | default "dagre" }}</li>
+  </ul>
+
+  {{ block "main" . }}
+  {{ end }}
 </body>
-<script src="/js/mermaid-init.js"></script>
-<script src="/js/mermaid.js"></script>
+{{- partial "assets/symbols.html" . -}}
+{{- partial "footer/scripts.html" (dict "page" .) -}}
+{{- partial "footer/scripts.html" (dict "page" . "match" "js/**.mjs" "type" "module" "template" true "filename" "js/mermaid.bundle.js") -}}
 </html>
