+++ go.mod
@@ -1,3 +1,5 @@
-module github.com/gethinode/mod-mermaid/v2
+module github.com/gethinode/mod-mermaid/v3
 
 go 1.19
+
+require github.com/gethinode/mod-fontawesome/v2 v2.1.2 // indirect
