+++ layouts/_shortcodes/mermaid.html
@@ -1,3 +1,26 @@
-<pre class="mermaid">
-    {{- .Inner | safeHTML -}}
-</pre>
+{{/* 
+    Copyright © 2025 The Hinode Team / Mark Dumay. All rights reserved.
+    Use of this source code is governed by The MIT License (MIT) that can be found in the LICENSE file.
+    Visit gethinode.com/license for more details.
+*/}}
+
+{{ $error := false }}
+
+{{/* Validate and initialize arguments */}}
+{{ $args := partial "utilities/InitArgs.html" (dict "structure" "mermaid" "args" .Params "named" .IsNamedParams "group" "shortcode") }}
+{{ if or $args.err $args.warnmsg }}
+    {{ partial (cond $args.err "utilities/LogErr.html" "utilities/LogWarn.html") (dict 
+        "partial"  "shortcodes/mermaid.html" 
+        "warnid"   "warn-invalid-arguments"
+        "msg"      "Invalid arguments"
+        "details"  ($args.errmsg | append $args.warnmsg)
+        "file"     page.File
+        "position" .Position
+    )}}
+    {{ $error = $args.err }}
+{{ end }}
+
+{{/* Main code */}}
+{{ if not $error }}
+    {{ partial "assets/mermaid.html" (dict "page" .Page "raw" .Inner "controls" $args.controls) }}
+{{ end }}
